requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
idna==2.10
joblib==1.0.1
mysql-connector-python==8.0.23
nltk==3.5
praw==7.1.4
prawcore==1.5.0
protobuf==3.14.0
regex==2020.11.13
requests==2.25.1
six==1.15.0
tqdm==4.57.0
update-checker==0.18.0
urllib3==1.26.3
websocket-client==0.57.0